Re-use dragHandler across normal, photo and icon calendar task elements
