+++ src/routes/[user]/components/Calendar/DayColumn.svelte
@@ -163,18 +163,11 @@
   {#each scheduledTasks as task, i (task.id)}
     <div class="task-absolute" style="top: {getOffset({ dt1: dt, dt2: getDateTimeFromTask(task) })}px;">
       {#if task.iconURL}
-        <IconTaskElement {task}
-          fontSize={0.8}
-        />
+        <IconTaskElement {task} fontSize={0.8} />
       {:else if task.imageDownloadURL}
-        <PhotoTaskElement {task}
-          fontSize={0.8}
-        />
+        <PhotoTaskElement {task} fontSize={0.8} />
       {:else}
-        <TaskElement {task}
-          fontSize={0.8}
-          hasCheckbox
-        />
+        <TaskElement {task} fontSize={0.8} hasCheckbox />
       {/if}
     </div>
   {/each}
